+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+#
+- Fix list endpoints
+- Add ability to create a image with labels
+- Add lots of missing filter variants and document them
+
 # 0.3.3
 - Fix return type of `Image::inspect`
 - Make api structs like `Container` thread safe